refactor(sheets): split GM config sheet into per-adapter implementations
InteractiveItemSheet becomes a system-agnostic dispatcher that never visually
renders. Its render() routes to a new adapter contract method,
renderConfigSheet(actor, options), so each adapter ships a config sheet that
matches the active system's UI conventions:

- Dnd5eInteractiveItemConfigSheet — ApplicationV2, mirrors the previous
  HandlebarsApplicationMixin(ActorSheetV2) implementation with PARTS,
  static actions, _prepareSubmitData, prose-mirror / file-picker custom
  elements.
- Pf2eInteractiveItemConfigSheet — ApplicationV1, matching pf2e's own
  AppV1 actor/item sheets: getData / _updateObject / activateListeners,
  _getHeaderButtons for the lock/identify/open-close toggles, file-picker
  buttons handled by FormApplication, and the {{editor}} Handlebars helper
  for rich-text fields.

Also adds isPhysicalItem(item) to the SystemAdapter contract so each system
defines its own physical-item predicate (dnd5e: "quantity" in system; pf2e:
PHYSICAL_ITEM_TYPES set). Both config sheets cache instances per actor UUID
so repeat opens reuse the same window.
