+++ templates/partials/config-fields-v1.hbs
@@ -1,0 +1,62 @@
+{{!-- Shared form fields for the pf2e (V1) config dialog. Uses V1 conventions:
+      file-picker buttons (handled by FormApplication._activateFilePicker) and
+      the {{editor}} Handlebars helper for rich text. --}}
+
+<div class="form-group">
+  <label>{{localize "INTERACTIVE_ITEMS.Sheet.Name"}}</label>
+  <div class="form-fields">
+    <input type="text" name="name" value="{{actor.name}}"
+      placeholder="{{localize 'INTERACTIVE_ITEMS.Sheet.NamePlaceholder'}}" />
+  </div>
+</div>
+
+<div class="form-group">
+  <label>{{localize "INTERACTIVE_ITEMS.Sheet.UnidentifiedName"}}</label>
+  <div class="form-fields">
+    <input type="text" name="system.unidentifiedName" value="{{system.unidentifiedName}}"
+      placeholder="{{localize 'INTERACTIVE_ITEMS.Sheet.UnidentifiedNamePlaceholder'}}" />
+  </div>
+</div>
+
+<div class="form-group">
+  <label>{{localize "INTERACTIVE_ITEMS.Sheet.LimitedName"}}</label>
+  <div class="form-fields">
+    <input type="text" name="system.limitedName" value="{{system.limitedName}}"
+      placeholder="{{localize 'INTERACTIVE_ITEMS.Sheet.LimitedNamePlaceholder'}}" />
+  </div>
+</div>
+
+<div class="form-group stacked">
+  <label>{{localize "INTERACTIVE_ITEMS.Sheet.Description"}}</label>
+  <div class="form-fields">
+    {{editor descriptionHTML target="system.description" button=true editable=isEditable engine="prosemirror"}}
+  </div>
+</div>
+
+<div class="form-group stacked">
+  <label>{{localize "INTERACTIVE_ITEMS.Sheet.UnidentifiedDescription"}}</label>
+  <div class="form-fields">
+    {{editor unidentifiedDescriptionHTML target="system.unidentifiedDescription" button=true editable=isEditable engine="prosemirror"}}
+  </div>
+</div>
+
+<div class="form-group stacked">
+  <label>{{localize "INTERACTIVE_ITEMS.Sheet.LimitedDescription"}}</label>
+  <div class="form-fields">
+    {{editor limitedDescriptionHTML target="system.limitedDescription" button=true editable=isEditable engine="prosemirror"}}
+  </div>
+</div>
+
+<div class="form-group">
+  <label>{{localize "INTERACTIVE_ITEMS.Sheet.InteractionRange"}}</label>
+  <div class="form-fields">
+    <input type="number" name="system.interactionRange" value="{{system.interactionRange}}" min="0" step="0.5" />
+  </div>
+</div>
+
+<div class="form-group">
+  <label>{{localize "INTERACTIVE_ITEMS.Sheet.InspectionRange"}}</label>
+  <div class="form-fields">
+    <input type="number" name="system.inspectionRange" value="{{system.inspectionRange}}" min="0" step="0.5" />
+  </div>
+</div>
